+++ Source/Krypton Components/Krypton.Toolkit/Controls Toolkit/KryptonForm.cs
@@ -246,7 +246,92 @@ namespace Krypton.Toolkit
         }
         #endregion
 
-        #region Public
+        #region Public (New)
+        /// <summary>
+        /// Toggles display of the minimize button.
+        /// </summary>
+        [DefaultValue(true)]
+        [Category("Window Style")]
+        [Description("Toggles display of the minimize button.")]
+        [DesignerSerializationVisibility(DesignerSerializationVisibility.Visible)]
+        public new bool MinimizeBox 
+        {
+            get => base.MinimizeBox;
+
+            set
+            {
+                if (base.MinimizeBox != value)
+                {
+                    base.MinimizeBox = value;
+                    _buttonManager.PerformNeedPaint(true);
+                }
+            }
+        }
+
+        /// <summary>
+        /// Toggles display of the maximize button.
+        /// </summary>
+        [DefaultValue(true)]
+        [Category("Window Style")]
+        [Description("Toggles display of the maximize button.")]
+        [DesignerSerializationVisibility(DesignerSerializationVisibility.Visible)]
+        public new bool MaximizeBox 
+        {
+            get => base.MaximizeBox;
+
+            set
+            {
+                if (base.MaximizeBox != value)
+                {
+                    base.MaximizeBox = value;
+                    _buttonManager.PerformNeedPaint(true);
+                }
+            }
+        }
+
+        /// <summary>
+        /// Toggles display of the Close button.
+        /// </summary>
+        [DefaultValue(true)]
+        [Category("Window Style")]
+        [Description("Toggles display of the close button.")]
+        [DesignerSerializationVisibility(DesignerSerializationVisibility.Visible)]
+        public new bool CloseBox 
+        {
+            get => base.CloseBox;
+
+            set
+            {
+                if (base.CloseBox != value)
+                {
+                    base.CloseBox = value;
+                    _buttonManager.PerformNeedPaint(true);
+                }
+            }
+        }
+
+        /// <summary>
+        /// Indicates the appearance and behavior of the border and title bar of the form.
+        /// </summary>
+        [Category("Appearance")]
+        [DefaultValue(FormBorderStyle.Sizable)]
+        [Description("Indicates the appearance and behavior of the border and title bar of the form.")]
+        public new FormBorderStyle FormBorderStyle 
+        {
+            get => base.FormBorderStyle;
+
+            set
+            {
+                if (base.FormBorderStyle != value)
+                {
+                    base.FormBorderStyle = value;
+                    OnFormBorderStyleChanged();
+                    _buttonManager.PerformNeedPaint(true);
+                }
+            }
+        }
+
+
         /// <summary>
         /// Gets or sets the extra text associated with the control.
         /// </summary>
@@ -1269,6 +1354,42 @@ namespace Krypton.Toolkit
         #endregion
 
         #region Implementation
+        private void OnFormBorderStyleChanged()
+        {
+            // KryptonForm uses ButtonSpecs for Form control buttons.
+            // Those need synchronizing when the FormBorderStyle changes.
+            // Once the style has change the user can adjust the buttons to the liking.
+            // User configured buttons are changed once the FormBorderStyle changes.
+
+            switch (FormBorderStyle)
+            {
+                case FormBorderStyle.None:
+                    ControlBox = false;
+                    MinimizeBox = false;
+                    MaximizeBox = false;
+                    CloseBox = false;
+                    break;
+
+                case FormBorderStyle.FixedSingle:
+                case FormBorderStyle.Fixed3D:
+                case FormBorderStyle.FixedDialog:
+                case FormBorderStyle.Sizable:
+                    ControlBox = true;
+                    MinimizeBox = true;
+                    MaximizeBox = true;
+                    CloseBox = true;
+                    break;
+
+                case FormBorderStyle.FixedToolWindow:
+                case FormBorderStyle.SizableToolWindow:
+                    ControlBox = false;
+                    MinimizeBox = false;
+                    MaximizeBox = false;
+                    CloseBox = true;
+                    break;
+            }
+        }
+
         private Icon? GetDefinedIcon()
         {
             // Are we allowed to try and show an icon?